requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
beautifulsoup4==4.6.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
Django==2.1.7
django-disqus==0.5
django-filter==2.0.0
django-modelcluster==4.2
django-react-templatetags==5.2.1
django-taggit==0.23.0
django-treebeard==4.3
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.3
djangorestframework==3.8.2
draftjs-exporter==2.1.5
elasticsearch==6.3.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
html5lib==1.0.1
idna==2.7
mock==2.0.0
mysqlclient==1.3.13
numpy==1.17.0
pbr==5.1.1
Pillow==5.2.0
promise==2.2.1
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.5
requests==2.20.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
Unidecode==1.0.22
urllib3==1.23
wagtail==2.4
webencodings==0.5.1
Willow==1.1
